+++ packages/backend/main.js
@@ -17,30 +17,8 @@ const provider = ethers.getDefaultProvider('matic');
 const signer = new ethers.Wallet(process.env.SEACRET_ADDRESS, provider);
 const pairContract = new ethers.Contract(address.QUICKSWAP, pairContractAbi, signer);
 
-// emitされたイベントに反応する
-const onMint = (senderAddress, amount0, amount1) => {
-    console.log("senderAddress: ", senderAddress);
-    console.log("amount0: ", amount0);
-    console.log("amount1: ", amount1);
-};
-
-const onBurn = (senderAddress, amount0, amount1, to) => {
-    console.log("senderAddress: ", senderAddress);
-    console.log("amount0: ", amount0);
-    console.log("amount1: ", amount1);
-    console.log("to: ", to);
-};
-
-const onSwap = (senderAddress, amount0In, amount1In, amount0Out, amount1Out, to) => {
-    console.log("senderAddress: ", senderAddress);
-    console.log("amount0In: ", amount0In, ", amount0Out: ", amount0Out);
-    console.log("amount1In: ", amount1In, ", amount1Out: ", amount1Out);
-    console.log("to: ", to);
-};
-
-pairContract.on("Mint", onMint);
-pairContract.on("Burn", onBurn);
-pairContract.on("Swap", onSwap);
+const quickswap = require('./quickswapEventHandler');
+quickswap.on(pairContract);
 
 // REST API
 const server = app.listen(3002, function(){
